fixed representation.py

+++ representation.py
@@ -14,14 +14,18 @@ class Representation:
     def __init__(self, path):
         dir = Path(path)
         if dir.is_dir():
-            self.dir = dir
-            self.nmeps = []
-            self.representations = {'raw': [], 'norm': [], 'tsfresh': [], 'tsfresh_fs': [], 'labels': []}
-            self.__read_data()
-            self.__extract_features()
-            self.__select_features()
-            with open('reps.pickle', 'wb') as f:
-                pickle.dump(self.representations, f)
+            reps = dir / 'reps.pickle'
+            if not reps.is_file():
+                self.dir = dir
+                self.nmeps = []
+                self.representations = {'raw': [], 'norm': [], 'tsfresh': [], 'tsfresh_fs': [], 'labels': []}
+                self.__read_data()
+                self.__extract_features()
+                self.__select_features()
+                with open(reps, 'wb') as f:
+                    pickle.dump(self.representations, f)
+            else:
+                self.representations = pd.read_pickle(reps)
         else:
             raise ValueError("input path not directory")
 
@@ -40,12 +44,12 @@ class Representation:
         for i, row in enumerate(m):
             d.append([[x, i, row[-1]] for x in row[:-1]])
         feat_ext = extract_features(timeseries_container=pd.DataFrame(np.vstack(d), columns=['value', 'id', 'kind']),
-                             column_id="id", column_kind="kind", column_value="value", n_jobs=4)
+                                    column_id="id", column_kind="kind", column_value="value", n_jobs=4)
         self.nmeps.append(0)
         [self.nmeps.append(len(x)) for x in self.representations['labels']]
         self.nmeps = np.cumsum(self.nmeps)
-        for idx in range(len(self.nmeps)-1):
-            self.representations['tsfresh'].append(feat_ext.iloc[self.nmeps[idx]:self.nmeps[idx+1], :])
+        for idx in range(len(self.nmeps) - 1):
+            self.representations['tsfresh'].append(feat_ext.iloc[self.nmeps[idx]:self.nmeps[idx + 1], :])
 
     def __select_features(self):
         m = np.append(np.vstack(self.representations['raw']),
@@ -54,7 +58,10 @@ class Representation:
         for i, row in enumerate(m):
             d.append([[x, i, row[-1]] for x in row[:-1]])
         y = pd.Series(np.hstack(self.representations['labels']))
-        feat_sel = extract_relevant_features(pd.DataFrame(np.vstack(d), columns=['value', 'id', 'kind']), y, column_id="id", column_kind="kind", column_value="value", n_jobs=4)
-        for idx in range(len(self.nmeps)-1):
-            self.representations['tsfresh_fs'].append(feat_sel.iloc[self.nmeps[idx]:self.nmeps[idx+1], :])
+        feat_sel = extract_relevant_features(pd.DataFrame(np.vstack(d), columns=['value', 'id', 'kind']), y,
+                                             column_id="id", column_kind="kind", column_value="value", n_jobs=4)
+        for idx in range(len(self.nmeps) - 1):
+            self.representations['tsfresh_fs'].append(feat_sel.iloc[self.nmeps[idx]:self.nmeps[idx + 1], :])
 
+    def get_representations(self):
+        return self.representations
